Improve version detection
Up until this, the third digit of the version register (patch verison)
was fetched from the number of git commits since the last tag instead
of parsing the tag itself. This commit moves the revision to bits 30:24
in the version register and fetches the patch version from the latest
tag.