Fix debugger access to syscon with EL2
This fixes a width mismatch in ID tags on the AXI bus when using EL2
which caused debug accesseses to hang the bus.

Note: Simply increasing the width by one causes a width mismatch when using EH1
but this is harmless and the extra logic will be optimized away. A potential
future fix would take the CPU type into consideration somehow.