Fix off by one error in SPI bootloader
s4 contains the number of bytes copied so far, and s0 contains the number of bytes to copy. When these are equal, we have copied everything and we shouldn't run one more iteration.

If we run the extra iteration and the flash image is being loaded so it lines up with the end of system memory, this causes a write past the end of system memory.